+++ src/llm_agents_from_scratch/subagents/__init__.py
@@ -1,0 +1,5 @@
+"""Subagents module."""
+
+from .spec import SubAgentSpec
+
+__all__ = ["SubAgentSpec"]